+++ src/commands/force/source/deploy.ts
@@ -6,7 +6,7 @@
  */
 import * as os from 'os';
 import { flags, FlagsConfig } from '@salesforce/command';
-import { Messages } from '@salesforce/core';
+import { Messages, SfdxError } from '@salesforce/core';
 import { AsyncResult, DeployResult } from '@salesforce/source-deploy-retrieve';
 import { Duration } from '@salesforce/kit';
 import { getString, isString } from '@salesforce/ts-types';
@@ -22,6 +22,9 @@ import { DeployProgressStatusFormatter } from '../../../formatters/deployProgres
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('@salesforce/plugin-source', 'deploy');
+
+// One of these flags must be specified for a valid deploy.
+const requiredFlags = ['manifest', 'metadata', 'sourcepath', 'validateddeployrequestid'];
 
 type TestLevel = 'NoTestRun' | 'RunSpecifiedTests' | 'RunLocalTests' | 'RunAllTestsInOrg';
 
@@ -109,6 +112,11 @@ export class Deploy extends DeployCommand {
   });
 
   public async run(): Promise<DeployCommandResult | DeployCommandAsyncResult> {
+    // verify that the user defined one of: manifest, metadata, sourcepath, validateddeployrequestid
+    if (!Object.keys(this.flags).some((flag) => requiredFlags.includes(flag))) {
+      throw SfdxError.create('@salesforce/plugin-source', 'deploy', 'MissingRequiredParam', [requiredFlags.join(', ')]);
+    }
+
     await this.deploy();
     this.resolveSuccess();
     return this.formatResult();